backend/requirements.txt
# Core dependencies
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
python-multipart==0.0.20

# OpenAI for transcription and GPT-5
openai==1.59.4

# Google Cloud Storage
google-cloud-storage==2.20.0
google-cloud-tasks==2.18.0

# Video processing
moviepy==2.1.1
ffmpeg-python==0.2.0

# PDF and PowerPoint processing
PyPDF2==3.1.0
pdf2image==1.17.0
python-pptx==1.0.2
Pillow==11.0.0

# Subtitle generation
pysrt==1.1.2

# API utilities
httpx==0.28.1
aiofiles==24.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development tools
pytest==8.3.4
pytest-asyncio==0.25.2
black==25.1.0
ruff==0.9.2

# Monitoring
prometheus-client==0.22.2
